+++ lib/icc_parser.dart
@@ -5,4 +5,7 @@ export 'src/cmm/color_profile_transform.dart' show ColorProfileTransform;
 export 'src/cmm/enums.dart';
 export 'src/color_profile.dart';
 export 'src/error.dart';
+export 'src/types/color_profile_header.dart';
+export 'src/types/color_profile_matrix.dart';
+export 'src/types/color_profile_primitives.dart';
 export 'src/utils/data_stream.dart';