+++ lib/src/cmm/color_profile_transform_4d.dart
@@ -173,7 +173,7 @@ final class ColorProfileTransform4DLut extends ColorProfileTransform {
     ColorProfileMatrix? usedMatrix;
     final matrix = tag.matrix;
     if (matrix != null) {
-      if (!matrix.isIdentity()) {
+      if (!matrix.isIdentity) {
         usedMatrix = matrix;
       }
     }