+++ test/cmm/color_profile_pcs_transform_test.dart
@@ -1,5 +1,8 @@
-import 'package:icc_parser/icc_parser.dart';
+import 'package:icc_parser/src/cmm/color_profile_cmm.dart';
 import 'package:icc_parser/src/cmm/color_profile_pcs_transform.dart';
+import 'package:icc_parser/src/cmm/color_profile_transform.dart';
+import 'package:icc_parser/src/color_profile.dart';
+import 'package:icc_parser/src/error.dart';
 import 'package:icc_parser/src/types/color_profile_header.dart';
 import 'package:icc_parser/src/types/color_profile_primitives.dart';
 import 'package:icc_parser/src/utils/list_utils.dart';